add `requirements.txt` to simplify dependency installation
this way dependencies can be added/removed in the future and consumers
only need to know that they need to run the standard `pip install -r
requirements.txt`, without having to care about which dependencies they
need in which version.

the file was created on Ubuntu 20.04 (LTS) with ROS 1 noetic using
`pipreqs --mode compat .`.