add initial `pyproject.toml`
this adds a [PEP 621] compatible [`pyproject.toml`].

this then allows running `pip install .` which would install the project
itself as well as its dependencies.
as the project doesn't have a `src/better_teleop` folder but instead
only contains `scripts/` nothing is installed for the project, but its
dependencies are nevertheless being installed.

the alternative to this would be to use a `requirements.txt` and then
require the users to run `pip install -r requirements.txt`. i've
prepared this in PR #1, but feel that this here is the better, more
modern and more extensible approach.

full disclosure: this is the first time i'm adding a `pyproject.toml` -
i'm not exactly familiar with python projects!

[PEP 621]: https://peps.python.org/pep-0621/
[`pyproject.toml`]: https://packaging.python.org/en/latest/specifications/declaring-project-metadata/